need to change git dispatch to use multiprocessing

+++ qa_backend/services/database/git_webhook_database.py
@@ -4,13 +4,14 @@ A GitDatabase which spawns a webhook listener.
 When the listener gets triggered, the database does a pull and fires an event
 """
 
-from typing import Optional, List, Callable, Coroutine, Any
+from typing import Optional, List, Callable, Coroutine, Any, MutableMapping
 from asyncio import Lock
 import asyncio
 import multiprocessing
 import os
 import signal
 
+from qa_backend import check_config_keys, ConfigurationError
 from .git_database import GitDatabase
 from qa_backend.server.webhook_listener import WebhookListener
 
@@ -34,6 +35,20 @@ class GitWebhookDatabase(GitDatabase):
         if start_listener:
             self.start_listener()
 
+    @staticmethod
+    def from_config(
+            config: MutableMapping[str,str]
+        ) -> 'GitWebhookDatabase':
+        check_config_keys(config, ['git dir', 'host', 'path', 'port'])
+        try:
+            git_dir = config['git dir']
+            host = str(config.get('host','0.0.0.0'))
+            port = int(config.get('port', 8082))
+            path = str(config.get('path','/webhook'))
+            return GitWebhookDatabase(git_dir,host=host,path=path,port=port)
+        except ValueError as e:
+            raise ConfigurationError(str(e))
+
     def add_callback(self, callback: Callable[[],Coroutine[Any,Any,None]]):
         self.flash_callbacks.append(callback)
 
@@ -44,14 +59,20 @@ class GitWebhookDatabase(GitDatabase):
 
     def kill_all_processes(self):
         for process in self.processes: 
-            os.kill(signal.SIGINT, process.pid)
-            process.terminate()
+            if process.is_alive():
+                os.kill(signal.SIGINT, process.pid)
+                process.terminate()
             process.join()
 
     def cancel_all_futures(self):
         # TODO: something smarter
         for future in self.futures:
             future.cancel()
+
+    async def shutdown(self):
+        self.kill_all_processes()
+        self.cancel_all_futures()
+        await asyncio.tasks.gather(*self.futures)
 
     async def flash_db(self):
         await self.reset()
